backend/requirements.txt
Django==3.2
psycopg2-binary==2.8.6
djangorestframework==3.12.4
drf-yasg==1.20.0
django-cors-headers==3.7.0
